API fixed

+++ src/utils/Api.js
@@ -1,4 +1,3 @@
-// Import BASE_URL from constants.js
 import { BASE_URL } from './constants';
 
 function checkResponse(res) {
@@ -10,11 +9,23 @@ function checkResponse(res) {
 }
 
 function getItems() {
-	return fetch(`${BASE_URL}/items`).then((res) => checkResponse(res));
+	const token = localStorage.getItem('token');
+	if (!token) {
+		return Promise.reject('No token available');
+	}
+
+	return fetch(`${BASE_URL}/items`, {
+		headers: {
+			Authorization: `Bearer ${token}`, // Add token to headers
+		},
+	}).then((res) => checkResponse(res));
 }
 
 function postItems(generatedData) {
 	const token = localStorage.getItem('token');
+	if (!token) {
+		return Promise.reject('No token available');
+	}
 
 	return fetch(`${BASE_URL}/items`, {
 		method: 'POST',
@@ -27,9 +38,12 @@ function postItems(generatedData) {
 }
 
 function deleteItem(id) {
-	if (id) {
-		const token = localStorage.getItem('token');
+	const token = localStorage.getItem('token');
+	if (!token) {
+		return Promise.reject('No token available');
+	}
 
+	if (id) {
 		return fetch(`${BASE_URL}/items/${id}`, {
 			method: 'DELETE',
 			headers: {
@@ -41,6 +55,9 @@ function deleteItem(id) {
 
 function likeItem(itemId) {
 	const token = localStorage.getItem('token');
+	if (!token) {
+		return Promise.reject('No token available');
+	}
 
 	return fetch(`${BASE_URL}/items/${itemId}/likes`, {
 		method: 'PUT',
@@ -52,6 +69,9 @@ function likeItem(itemId) {
 
 function dislikeItem(itemId) {
 	const token = localStorage.getItem('token');
+	if (!token) {
+		return Promise.reject('No token available');
+	}
 
 	return fetch(`${BASE_URL}/items/${itemId}/likes`, {
 		method: 'DELETE',
